Support conversin to MP4 using ffmpeg
Adds an option to convert to MP4 for smaller file sizes. This introduces
a runtime dependencies on ffmpeg-tools.

A new option --usedconf has also been added to the background daemon,
which is activated by default. When set the daemon will read in the
configuration values from dconf every time a new recording is started,
ensuring that the action is consistent with the values set through the
UI.